Moved Connectivity code to network module
