+++ parley/src/main/java/nu/parley/android/data/net/Connectivity.java
@@ -30,72 +30,6 @@ public final class Connectivity {
     private static final String HEADER_PARLEY_IDENTIFICATION = "x-iris-identification"; // Rename Iris to Parley when backend accepts this
     private static final String HEADER_PARLEY_AUTHORIZATION = "Authorization";
 
-    /**
-     * Get a Retrofit instance with Parley's url and path and OkHttpClient.
-     *
-     * @return Retrofit
-     */
-    public static Retrofit getRetrofit() {
-        Retrofit.Builder retrofitBuilder = new Retrofit.Builder()
-                .baseUrl(Parley.getInstance().getNetwork().getBaseUrl())
-                .addConverterFactory(GsonConverterFactory.create())
-                .client(getOkHttpClient());
-
-        return retrofitBuilder.build();
-    }
-
-    /**
-     * Get a OkHttpClient instance with Parley identification header.
-     *
-     * @return OkHttpClient
-     */
-    private static OkHttpClient getOkHttpClient() {
-        OkHttpClient.Builder okHttpClientBuilder = new OkHttpClient.Builder()
-                .addInterceptor(new Interceptor() {
-                    @Override
-                    public Response intercept(Chain chain) throws IOException {
-                        Request original = chain.request();
-
-                        Request.Builder requestBuilder = original.newBuilder()
-                                .method(original.method(), original.body());
-
-                        for (Map.Entry<String, String> entry : getAdditionalHeaders().entrySet()) {
-                            requestBuilder.addHeader(entry.getKey(), entry.getValue());
-                        }
-                        for (Map.Entry<String, String> entry : getParleyHeaders().entrySet()) {
-                            requestBuilder.addHeader(entry.getKey(), entry.getValue());
-                        }
-
-                        Request request = requestBuilder.build();
-                        return chain.proceed(request);
-                    }
-                })
-                .connectTimeout(30, TimeUnit.SECONDS)
-                .readTimeout(30, TimeUnit.SECONDS)
-                .writeTimeout(30, TimeUnit.SECONDS);
-
-        okHttpClientBuilder = addInterceptor(okHttpClientBuilder);
-
-        applySslPinning(okHttpClientBuilder);
-
-        return okHttpClientBuilder.build();
-    }
-
-    private static void applySslPinning(OkHttpClient.Builder okHttpClientBuilder) {
-        URL url;
-        String serverHostname;
-        try {
-            url = new URL(Parley.getInstance().getNetwork().url);
-            serverHostname = url.getHost();
-
-            SSLSocketFactory sslSocketFactory = TrustKit.getInstance().getSSLSocketFactory(serverHostname);
-            X509TrustManager trustManager = TrustKit.getInstance().getTrustManager(serverHostname);
-            okHttpClientBuilder.sslSocketFactory(sslSocketFactory, trustManager);
-        } catch (MalformedURLException e) {
-            e.printStackTrace();
-        }
-    }
-
     public static Map<String, String> getAdditionalHeaders() {
         return Parley.getInstance().getNetwork().headers;
     }
@@ -107,19 +41,6 @@ public final class Connectivity {
             headers.put(HEADER_PARLEY_AUTHORIZATION, Parley.getInstance().getUserAuthorization());
         }
         return headers;
-    }
-
-    /**
-     * Adds an interceptor to the OkHttpClient.Builder.
-     *
-     * @param builder the OkHttpClient.Builder to add the interceptor to.
-     * @return the OkHttpClient.Builder with the added interceptor.
-     */
-    private static OkHttpClient.Builder addInterceptor(OkHttpClient.Builder builder) {
-        if (Parley.getInstance().getNetwork().interceptor != null) {
-            return builder.addInterceptor(Parley.getInstance().getNetwork().interceptor);
-        }
-        return builder;
     }
 
     public static String toMediaUrl(String mediaId) {
@@ -135,22 +56,5 @@ public final class Connectivity {
             lazyHeadersBuilder.addHeader(entry.getKey(), entry.getValue());
         }
         return new GlideUrl(url, lazyHeadersBuilder.build());
-    }
-
-    public static <T> String getFormattedError(retrofit2.Response<T> response) {
-        try {
-            if (response.errorBody() == null) {
-                return response.message();
-            }
-            ParleyErrorResponse error = new Gson().fromJson(response.errorBody().string(), ParleyErrorResponse.class);
-            String message = error.getMessage();
-            if (message == null) {
-                return response.message();
-            } else {
-                return message;
-            }
-        } catch (JsonSyntaxException | IOException e) {
-            return response.message();
-        }
     }
 }
