+++ parley/src/main/java/nu/parley/android/data/net/response/ParleyResponsePostMedia.java
@@ -6,7 +6,6 @@ import com.google.gson.annotations.SerializedName;
 
 public final class ParleyResponsePostMedia {
 
-    @SerializedName("media")
     @Nullable
     public final String mediaId;
