+++ parley/src/main/java/nu/parley/android/data/net/response/ParleyResponsePostMessage.java
@@ -6,7 +6,6 @@ import com.google.gson.annotations.SerializedName;
 
 public final class ParleyResponsePostMessage {
 
-    @SerializedName("messageId")
     @Nullable
     private String messageId;
